requirements.txt
django
statsmodels
pandas
python-dotenv
requests
Flask
gunicorn
django-heroku
